+++ src/lib.rs
@@ -1,3 +1,5 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
 //! # High-order Virtual Machine (HVM) library
 //!
 //! Note: this API is **unstable**.
@@ -5,6 +7,7 @@
 // FIXME: what is the right way to export the definitions on api.rs as a lib?
 
 pub mod builder;
+#[cfg(feature = "std")]
 pub mod compiler;
 pub mod language;
 pub mod parser;
@@ -12,6 +15,10 @@ pub mod readback;
 pub mod rulebook;
 pub mod runtime;
 pub mod api;
+
+extern crate alloc;
+
+use alloc::string::{String, ToString};
 
 pub use builder::eval_code;
