+++ src/main.rs
@@ -1,5 +1,10 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
 mod api;
 mod builder;
+#[cfg(feature = "std")]
+mod cli;
+#[cfg(feature = "std")]
 mod compiler;
 mod language;
 mod parser;
@@ -7,74 +12,16 @@ mod readback;
 mod rulebook;
 mod runtime;
 
-use clap::{Parser, Subcommand};
-
-#[derive(Parser)]
-#[clap(author, version, about, long_about = None)]
-#[clap(propagate_version = true)]
-pub struct Cli {
-  #[clap(subcommand)]
-  pub command: Command,
-}
-
-#[derive(Subcommand)]
-pub enum Command {
-  #[clap(about = "Run a file interpreted", name = "run", aliases = &["r"])]
-  Run { file: String, params: Vec<String> },
-
-  #[clap(about = "Compile file to C",  name = "compile", aliases = &["c"])]
-  Compile {
-    file: String,
-    #[clap(long)]
-    single_thread: bool,
-  },
-
-  #[clap(about = "Run in debug mode", name = "debug", aliases = &["d"])]
-  Debug { file: String, params: Vec<String> },
-}
+extern crate alloc;
 
 fn main() {
-  match run_cli() {
+  #[cfg(feature = "std")]
+  match cli::run_cli() {
     Ok(..) => {}
     Err(err) => {
       eprintln!("{}", err);
     }
   };
-}
-
-fn run_cli() -> Result<(), String> {
-  let cli_matches = Cli::parse();
-
-  fn hvm(file: &str) -> String {
-    if file.ends_with(".hvm") {
-      file.to_string()
-    } else {
-      format!("{}.hvm", file)
-    }
-  }
-
-  match cli_matches.command {
-    Command::Compile { file, single_thread } => {
-      let file = &hvm(&file);
-      let code = load_file_code(file)?;
-
-      compile_code(&code, file, !single_thread)?;
-      Ok(())
-    }
-    Command::Run { file, params } => {
-      let code = load_file_code(&hvm(&file))?;
-
-      run_code(&code, false, params)?;
-      Ok(())
-    }
-
-    Command::Debug { file, params } => {
-      let code = load_file_code(&hvm(&file))?;
-
-      run_code(&code, true, params)?;
-      Ok(())
-    }
-  }
 }
 
 fn make_call(params: &Vec<String>) -> Result<language::Term, String> {
@@ -97,6 +44,7 @@ fn run_code(code: &str, debug: bool, params: Vec<String>) -> Result<(), String> 
   Ok(())
 }
 
+#[cfg(feature = "std")]
 fn compile_code(code: &str, name: &str, parallel: bool) -> Result<(), String> {
   if !name.ends_with(".hvm") {
     return Err("Input file must end with .hvm.".to_string());
@@ -111,6 +59,7 @@ fn load_file_code(file_name: &str) -> Result<String, String> {
   std::fs::read_to_string(file_name).map_err(|err| err.to_string())
 }
 
+#[cfg(feature = "std")]
 #[allow(dead_code)]
 fn run_example() -> Result<(), String> {
   // Source code
